Refine MCP E2E fixture boundaries
Split read-only MCP SQL and execute_query E2E scenarios onto shared fixtures.
Narrow lowercase HTTP header coverage to the header contract and keep broad model contracts in dedicated tests.
Deduplicate production Proxy approval coverage and add a programmatic invisible-approved-step safety check.
