+++ test/e2e/mcp/src/test/java/org/apache/shardingsphere/test/e2e/mcp/runtime/programmatic/ExecuteQueryTransactionE2ETest.java
@@ -38,31 +38,6 @@ class ExecuteQueryTransactionE2ETest extends AbstractHttpProgrammaticRuntimeE2ET
     }
     
     @Test
-    void assertExecuteSelectOverHttpSession() throws IOException, InterruptedException {
-        launchHttpTransport();
-        HttpClient httpClient = HttpClient.newHttpClient();
-        String sessionId = initializeSession(httpClient);
-        HttpResponse<String> actual = sendToolCallRequest(httpClient, sessionId, "database_gateway_execute_query",
-                createExecuteQueryArguments("logic_db", "logic_db", "SELECT * FROM orders", 10));
-        assertThat(actual.statusCode(), is(200));
-        Map<String, Object> payload = getStructuredContent(actual.body());
-        assertThat(String.valueOf(payload.get("result_kind")), is("result_set"));
-    }
-    
-    @Test
-    void assertExecuteReadOnlyCommonTableExpression() throws IOException, InterruptedException {
-        launchHttpTransport();
-        HttpClient httpClient = HttpClient.newHttpClient();
-        String sessionId = initializeSession(httpClient);
-        HttpResponse<String> actual = sendToolCallRequest(httpClient, sessionId, "database_gateway_execute_query",
-                createExecuteQueryArguments("logic_db", "logic_db", "WITH foo_orders AS (SELECT * FROM orders) SELECT * FROM foo_orders"));
-        assertThat(actual.statusCode(), is(200));
-        Map<String, Object> payload = getStructuredContent(actual.body());
-        assertThat(String.valueOf(payload.get("result_kind")), is("result_set"));
-        assertThat(String.valueOf(payload.get("statement_class")), is("query"));
-    }
-    
-    @Test
     void assertRejectCrossDatabaseTransactionSwitch() throws IOException, InterruptedException {
         launchHttpTransport();
         HttpClient httpClient = HttpClient.newHttpClient();
@@ -101,7 +76,7 @@ class ExecuteQueryTransactionE2ETest extends AbstractHttpProgrammaticRuntimeE2ET
         assertThat(sendDeleteRequest(httpClient, createSessionHeaders(sessionId)).statusCode(), is(200));
         String newSessionId = initializeSession(httpClient);
         HttpResponse<String> actual = sendToolCallRequest(httpClient, newSessionId, "database_gateway_execute_query",
-                createExecuteQueryArguments("logic_db", "logic_db", "SELECT status FROM orders WHERE order_id = 1"));
+                Map.of("database", "logic_db", "schema", "logic_db", "sql", "SELECT status FROM orders WHERE order_id = 1"));
         assertThat(actual.statusCode(), is(200));
         Map<String, Object> payload = getStructuredContent(actual.body());
         assertThat(String.valueOf(((List<?>) ((List<?>) payload.get("rows")).get(0)).get(0)), is("NEW"));
@@ -118,13 +93,4 @@ class ExecuteQueryTransactionE2ETest extends AbstractHttpProgrammaticRuntimeE2ET
         assertThat(String.valueOf(getStructuredContent(actual.body()).get("message")), is(expectedMessage));
     }
     
-    private Map<String, Object> createExecuteQueryArguments(final String databaseName, final String schemaName, final String sql) {
-        return createExecuteQueryArguments(databaseName, schemaName, sql, -1);
-    }
-    
-    private Map<String, Object> createExecuteQueryArguments(final String databaseName, final String schemaName, final String sql, final int maxRows) {
-        return -1 == maxRows
-                ? Map.of("database", databaseName, "schema", schemaName, "sql", sql)
-                : Map.of("database", databaseName, "schema", schemaName, "sql", sql, "max_rows", maxRows);
-    }
 }
